report procedyre adjusted

+++ src/messages/ReportMessage.java
@@ -9,9 +9,9 @@ import peersim.core.Node;
 public class ReportMessage {
     public Node sender;
     public Node target;
-    public int  weigth;
+    public long  weigth;  // d(s,t) distance between sender and t
 
-    public ReportMessage(Node s, Node t, int w){
+    public ReportMessage(Node s, Node t, long w){
         sender= s;
         target = t;
         weigth = w;
